Turn test runners into test loggers.
Remove the responsibility of running the tests from TextTestRunner,
that is, make TextTestRunner just a logger of the results collected by
TestResult. Instead, use an autorun() handler to set up the required objects
and to execute the tests.

Given the recently introduced ability of adding more than one observer of
a TestResult object, we are now ready to implement different loggers, and
possibly running more than one logger at a time.