Fix problems with autosaving test results.
Update AppleScriptEditorLogger to fix some errors
produced by AppleScript Editor with creating a
document for the test results ("cannot autosave,
document doesn't exist") in Mavericks.

The way I fixed this is a bit convoluted, but I
haven't found a simpler way. All the pieces seem
necessary though (make sure the file exists with
"open access", use a POSIX path for the path
property when creating the new document, explicitly
save the document, and use an alias when saving).
I have also prefixed the output lines with -- so
they appear to be comments to the editor: this
allows the user to save the test window without
getting compiler errors.