Fix output to editor without breaking the console.
Use linefeed ("\n") instead of return ("\r"). The
latter does not start a new line in the Terminal.

Make sure that each line in the AS Editor's output
is prefixed with "--", but leave just empty lines
(without "--") at the end of the test report.

Use print[Colored]Line() only for whole lines, and
not for appending to the current line.