Add a load script test case in the template.
It is better to load the script to be tested inside a test case,
because this strategy works when the tests are loaded using
loadTestsFromFolder() (when the test scripts are loaded through
loadTestsFromFolder(), their run handler is not executed).
The alternative would be to load the script at compile time, but
that may be inconvenient.