Remove Minitest.
MiniTest is an interesting experiment, but its
advantages are not overwhelming (especially
starting with AppleScript 2.3) and it has some
drawbacks (running all tests from a folder does
not produce a single report). Besides, keeping two
frameworks that provide mostly the same
functionality is a burden for the developer and
may be confusing for users.