Make pp() exception-safe.
Modify the logic of pretty-printing so that an
exception is never raised. This simplified client
code, which can assume that any object can be
pretty-printed.

Remove the msg parameter from assertKindOf() and
related methods, and rely on printing a message
automatically instead.

Optimize the logic of [assert|refute]Equal() and
fix a bug causing these assertions to raise an error
when the class of the arguments could not be obtained.

Implement refuteKindOf() explicitly instead of
calling failIf().