Set the name property of top level script.
As of Mavericks, each library script should define
name, id and version properties. Explicitly setting
the name property at the top level is a bit tricky,
because the name property is inherited by all the
nested scripts that do not explicitly define their
parent to something different from the top-level
script.

Since ASUnit uses the names of the scripts for the
output, we must ensure that TestComponent does not
inherit from the top-level script (otherwise, all
the test names becomes “ASUnit”). Fortunately,
setting the parent of TestComponent to something
different (AppleScript) requires only a small change
(makeTestResult() is not inherited any more, so
we must target TOP_LEVEL).