Consistently use `my` to refer to own properties.
To correctly support the overriding of properties,
refer to each property `p` with the syntax `my p`.
Consider this example:

    script Father
      property x : 0
      on f()
        x
      end
    end

    script Child
      property parent : Father
      property x : 1
    end

    Child's f() -- Returns 0

Property `x` has been overridden, but not handler `f()`. Therefore,
`Child's f()` uses `Father's x` and not `Child's x`. To fix this misbehavior,
`Father` must be implemented as follows:

    script Father
      property x : 0
      on f()
        my x
      end f
    end

See also the AppleScript Language Guide > Inheritance of Script Objects >
Some Examples of Inheritance. In particular:

>The previous example demonstrates an important point about inherited properties:
>to refer to an inherited property from within a child script object, you must
>use the reserved word `my` or `of me` to indicate that the value to which you’re
>referring is a property of the current script object. (You can also use the words
>of parent to indicate that the value is a property of the parent script object.)
>If you don’t, AppleScript assumes the value is a local variable.

The exceptions are global constants like TEST_FAILED, whose value
is not supposed to be overridden.