Simplify AppleScriptEditorLogger's initialize().
The workaround I had put in place to get the
AppleScriptEditorLogger work with the autosave feature
in Mavericks can be simplified.

It turns out that the problem is that
the name property doesn't seem to play well with
autosaving. So, we can simply avoid setting any
properties of the new document.

See also: http://macscripter.net/viewtopic.php?id=42072