Update api task to use a local patched version of HeaderDoc.
This is a temporary change. Hopefully, the next
version of Xcode will ship with the patched HeaderDoc.

To build the documentation, you need to download
HeaderDoc 8.9.14 from

    http://opensource.apple.com/release/developer-tools-46/

and the patches headerdoc-8.9.14-8.9.24.diff,
headerdoc-8.9.24-8.9.25.diff, and headerdoc-8.9.25-8.9.26.diff
from

    http://www.darwin-development.org/headerdoc_patches/Xcode_5/

Download the patches inside the HeaderDoc main folder
and apply them with

    patch -p1 < headerdoc-8.9.14-8.9.24.diff

and similarly for the other two. Then, `make`.

The api task now assumes that the headerdoc-8.9.14 folder
is inside your ~/Downloads folder.