Avoid UTF-16 encoding in source files.
Script Editor automatically chooses the encoding of
a file upon saving it, based on the content of the
file. Scripts using only ASCII characters are saved
as ASCII. If there are non-ASCII characters such as
≠, «, », etc… Script Editor usually uses the MacOS Roman
encoding. But when the script cannot be represented
in that encoding, Script Editor saves it as UTF-16.

To make it easier to work with version control programs,
diff programs and so on, I prefer to avoid UTF-16.
To do so, some incriminated characters have been removed
(in the test file), and the remaining (used to show invisible
characters) have been replaced with «data» expressions.

See also: http://lists.apple.com/archives/applescript-users/2003/Jan/msg00741.html.