Improve pretty printing of objects of custom classes.
If an object has a custom class, such class object
may not be coercible to text. Consider, for instance:

    script
        property class : me — Weird, but legal
    end

Before this commit, pp() would just print «object»
for an object like the script above. With this commit,
the textual representation is slightly better (now
you get «object of class self» for the case above).

Note that the recursive call that we have introduced
might cause pp() to loop indefinitely (this must be
tested).