Use UTF-8 for the API documentation.
Although the source code is saved by Script Editor
as MacRoman, HeaderDoc produces better output with
UTF-8 (also, somehow when viewing the documentation
from GitHub, the encoding is not correct).

Of course, since there is a mismatch between
the input and the output encoding, care must be
taken in choosing the characters. In particular,
smart quotes and en-dashes must be replaced by
normal quotes and simple dashes, respectively.